move extension recommendations
